no bootstrap for you 10/12
Bootstrap was causing draw errors and it seemed best to just clear it from the package list. Presently it is commented out but it is still there. Also I added the api window to the view on the main page and created a db connect for working locally... make sure that you change the db connect file's values to set the correct connection script.