click function now changes colors and other neat stuff
I have managed to not only change the colors of the models when clicked but also load a new API instance in the place of the current API with a different model if the piece labeled with the 4 annotation next to it is clicked, not the annotation but the model itself. This means we can now bring in a different model with all of the pieces that belong to that section and load it in the place of the current model.